Split unit tests from IngestRecordsFromIteratorIT into new test file
